refactor: Create block-type-utils with block type constants
